* Fix a number of warnings reported by compilers on edge * Adjust regression test to tag make output with day of week information
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@24077 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/databases/Uintah/avtUintahFileFormat.C
@@ -812,7 +812,7 @@ avtUintahFileFormat::ReadMetaData(avtDatabaseMetaData *md, int timeState)
 
   cycles.resize( cycleTimes.size() );
 
-  for(int i=0; i<cycleTimes.size(); ++i )
+  for(int i=0; i<(int)cycleTimes.size(); ++i )
     cycles[i] = i;
 
   md->SetCycles( cycles );
